Add optional event registration to Drift provider component
The Drift provider component will now register drift event handlers
provided by the user. For now, only a few events are supported with
types, more to come.

There is no checking if the event name is valid, or that
the associated handler is a function as the drift script will log on both of these
cases and we don't want to cause a runtime error regardless.

Ideally the user will define their events object outside of a component
so that the object identity does not change on rerender. We ensure that only
one listener is attached per event in the case this identity does change. This is
also in place to allow for the case where certain events should be tracked in certain
situations.